Revision de codigo

+++ RateTheAnime-main/RateTheAnime-main/ratetheanime/src/main/java/com/duoc/ratetheanime/repository/UsuarioRepository.java
@@ -1,10 +1,11 @@
 package com.duoc.ratetheanime.repository;
 
-import com.duoc.ratetheanime.model.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.duoc.ratetheanime.model.Usuario;
+
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
     
 }
